backend/requirements.txt
Django==1.8.5
django-filter==0.11.0
django-grappelli==2.7.1
djangorestframework==3.3.1
Markdown==2.6.4
psycopg2==2.6.1
wheel==0.24.0
